Enforced coding style on project

+++ Buffs/LivingTombstoneBuff.cs
@@ -25,7 +25,7 @@ namespace Tremor.Buffs
 			}
 			if (petProjectileNotSpawned && player.whoAmI == Main.myPlayer)
 			{
-				Projectile.NewProjectile(player.position.X + (float)(player.width / 2), player.position.Y + (float)(player.height / 2), 0f, 0f, mod.ProjectileType("LivingTombstonePro"), 0, 0f, player.whoAmI, 0f, 0f);
+				Projectile.NewProjectile(player.position.X + player.width / 2, player.position.Y + player.height / 2, 0f, 0f, mod.ProjectileType("LivingTombstonePro"), 0, 0f, player.whoAmI, 0f, 0f);
 			}
 		}
 	}
