+++ Ice/Mobs/Frostbyte.cs
@@ -1,6 +1,6 @@
-using Microsoft.Xna.Framework;
 using System;
 using System.Linq;
+using Microsoft.Xna.Framework;
 using Terraria;
 using Terraria.ModLoader;
 
@@ -31,8 +31,8 @@ namespace Tremor.Ice.Mobs
 
 		public override void ScaleExpertStats(int numPlayers, float bossLifeScale)
 		{
-			npc.lifeMax = (int)(npc.lifeMax * 1);
-			npc.damage = (int)(npc.damage * 1);
+			npc.lifeMax = npc.lifeMax * 1;
+			npc.damage = npc.damage * 1;
 		}
 
 		public override float SpawnChance(NPCSpawnInfo spawnInfo)
@@ -50,8 +50,8 @@ namespace Tremor.Ice.Mobs
 		public override void AI()
 		{
 			int tileX = (int)(npc.Bottom.Y / 16f), tileY = (int)(npc.Bottom.Y / 16f);
-			int height = (int)Math.Min(10, tileX);
-			float velX = MathHelper.Lerp(5f, 3f, (float)height / 10f), velY = MathHelper.Lerp(-3f, -6f, (float)height / 10f);
+			int height = Math.Min(10, tileX);
+			float velX = MathHelper.Lerp(5f, 3f, height / 10f), velY = MathHelper.Lerp(-3f, -6f, height / 10f);
 			npc.aiStyle = 1;
 			if (npc.target >= 0 && npc.target <= 255 && Main.player[npc.target].Bottom.Y < npc.Center.Y && npc.collideX && npc.velocity.Y != 0)
 			{
@@ -61,7 +61,7 @@ namespace Tremor.Ice.Mobs
 					for (int m = 0; m < 2; m++)
 					{
 						float xPos = npc.velocity.X > 0 ? npc.Right.X : npc.Left.X;
-						int dustID = Dust.NewDust(new Vector2(xPos, npc.Center.Y), 1, 1, 80, xPos < npc.Center.X ? -4f : 4f, (float)Math.Abs(npc.velocity.Y) * 0.2f, 100, Color.White, 1.5f);
+						int dustID = Dust.NewDust(new Vector2(xPos, npc.Center.Y), 1, 1, 80, xPos < npc.Center.X ? -4f : 4f, Math.Abs(npc.velocity.Y) * 0.2f, 100, Color.White, 1.5f);
 						Main.dust[dustID].noGravity = true;
 					}
 				}
