+++ Invasion/CyberLaserBat.cs
@@ -28,7 +28,7 @@ namespace Tremor.Invasion
 			Vector2 lookTarget = projectile.Center + projectile.velocity;
 			float rotX = lookTarget.X - center.X;
 			float rotY = lookTarget.Y - center.Y;
-			projectile.rotation = -((float)Math.Atan2((double)rotX, (double)rotY)) - 1.57f;
+			projectile.rotation = -((float)Math.Atan2(rotX, rotY)) - 1.57f;
 			if (Main.netMode != 2)
 			{
 				int dustID = Dust.NewDust(projectile.position, projectile.width, projectile.height, mod.DustType("CyberDust"), projectile.velocity.X * 0.2f, projectile.velocity.Y * 0.2f, 100, Color.White, 0.6f);
