+++ Invasion/CyberWrathInvasion.cs
@@ -4,6 +4,7 @@ using Terraria;
 using Terraria.DataStructures;
 using Terraria.ModLoader;
 using Tremor.Invasion;
+
 //using Terraria.Content.Fonts;
 
 namespace Tremor
@@ -43,8 +44,8 @@ namespace Tremor
 			int eH = CyberWrathI.Height;
 			int eW = CyberWrathI.Width;
 
-			int XX1 = (int)(((24 - iW1) / 2) + Main.screenWidth - sX1);
-			int YY1 = (int)(((24 - iH1) / 2) + sY1 + (int)(280 * 1.4) + (24 - iW1) * (-1) + 20);
+			int XX1 = ((24 - iW1) / 2) + Main.screenWidth - sX1;
+			int YY1 = ((24 - iH1) / 2) + sY1 + (int)(280 * 1.4) + (24 - iW1) * (-1) + 20;
 
 			int eX = XX1 - 333;
 			int eY = YY1 - 430 + (24 - eW) * (-1);
