+++ Items/RedStorm.cs
@@ -1,7 +1,7 @@
-using Terraria.ID;
 using System;
 using Microsoft.Xna.Framework;
 using Terraria;
+using Terraria.ID;
 using Terraria.ModLoader;
 
 namespace Tremor.Items
@@ -54,19 +54,19 @@ namespace Tremor.Items
 			num74 = player.GetWeaponKnockback(item, num74);
 			player.itemTime = item.useTime;
 			Vector2 vector2 = player.RotatedRelativePoint(player.MountedCenter, true);
-			Vector2 value = Vector2.UnitX.RotatedBy((double)player.fullRotation, default(Vector2));
+			Vector2 value = Vector2.UnitX.RotatedBy(player.fullRotation, default(Vector2));
 			Vector2 vector3 = Main.MouseWorld - vector2;
-			float num78 = (float)Main.mouseX + Main.screenPosition.X - vector2.X;
-			float num79 = (float)Main.mouseY + Main.screenPosition.Y - vector2.Y;
+			float num78 = Main.mouseX + Main.screenPosition.X - vector2.X;
+			float num79 = Main.mouseY + Main.screenPosition.Y - vector2.Y;
 			if (player.gravDir == -1f)
 			{
-				num79 = Main.screenPosition.Y + (float)Main.screenHeight - (float)Main.mouseY - vector2.Y;
+				num79 = Main.screenPosition.Y + Main.screenHeight - Main.mouseY - vector2.Y;
 			}
-			float num80 = (float)Math.Sqrt((double)(num78 * num78 + num79 * num79));
+			float num80 = (float)Math.Sqrt(num78 * num78 + num79 * num79);
 			float num81 = num80;
 			if ((float.IsNaN(num78) && float.IsNaN(num79)) || (num78 == 0f && num79 == 0f))
 			{
-				num78 = (float)player.direction;
+				num78 = player.direction;
 				num79 = 0f;
 				num80 = num72;
 			}
@@ -79,11 +79,11 @@ namespace Tremor.Items
 			int num107 = 3;
 			for (int num108 = 0; num108 < num107; num108++)
 			{
-				vector2 = new Vector2(player.position.X + (float)player.width * 0.5f + (float)(Main.rand.Next(201) * -(float)player.direction) + ((float)Main.mouseX + Main.screenPosition.X - player.position.X), player.MountedCenter.Y - 600f);
-				vector2.X = (vector2.X + player.Center.X) / 2f + (float)Main.rand.Next(-200, 201);
-				vector2.Y -= (float)(100 * num108);
-				num78 = (float)Main.mouseX + Main.screenPosition.X - vector2.X;
-				num79 = (float)Main.mouseY + Main.screenPosition.Y - vector2.Y;
+				vector2 = new Vector2(player.position.X + player.width * 0.5f + Main.rand.Next(201) * -(float)player.direction + (Main.mouseX + Main.screenPosition.X - player.position.X), player.MountedCenter.Y - 600f);
+				vector2.X = (vector2.X + player.Center.X) / 2f + Main.rand.Next(-200, 201);
+				vector2.Y -= 100 * num108;
+				num78 = Main.mouseX + Main.screenPosition.X - vector2.X;
+				num79 = Main.mouseY + Main.screenPosition.Y - vector2.Y;
 				if (num79 < 0f)
 				{
 					num79 *= -1f;
@@ -92,15 +92,15 @@ namespace Tremor.Items
 				{
 					num79 = 20f;
 				}
-				num80 = (float)Math.Sqrt((double)(num78 * num78 + num79 * num79));
+				num80 = (float)Math.Sqrt(num78 * num78 + num79 * num79);
 				num80 = num72 / num80;
 				num78 *= num80;
 				num79 *= num80;
-				float speedX4 = num78 + (float)Main.rand.Next(-40, 41) * 0.02f;
-				float speedY5 = num79 + (float)Main.rand.Next(-40, 41) * 0.02f;
-				Projectile.NewProjectile(vector2.X, vector2.Y, speedX4, speedY5, 606, num73, num74, i, 0f, (float)Main.rand.Next(5));
-				Projectile.NewProjectile(vector2.X, vector2.Y, speedX4, speedY5, 606, num73, num74, i, 0f, (float)Main.rand.Next(3));
-				Projectile.NewProjectile(vector2.X, vector2.Y, speedX4, speedY5, 606, num73, num74, i, 0f, (float)Main.rand.Next(1));
+				float speedX4 = num78 + Main.rand.Next(-40, 41) * 0.02f;
+				float speedY5 = num79 + Main.rand.Next(-40, 41) * 0.02f;
+				Projectile.NewProjectile(vector2.X, vector2.Y, speedX4, speedY5, 606, num73, num74, i, 0f, Main.rand.Next(5));
+				Projectile.NewProjectile(vector2.X, vector2.Y, speedX4, speedY5, 606, num73, num74, i, 0f, Main.rand.Next(3));
+				Projectile.NewProjectile(vector2.X, vector2.Y, speedX4, speedY5, 606, num73, num74, i, 0f, Main.rand.Next(1));
 			}
 			return false;
 		}
